Some rename and factory cleanup
