+++ examples/data-objects/monaco/src/dataObject.ts
